requirements.txt
langchain==0.1.16
langchain-chroma
langchain-openai==0.1.3
langgraph==0.0.38
grandalf==0.8
python-dotenv==1.0.1
langsmith
chromadb
pypdf